backend/requirements.txt
django>=5.0
djangorestframework
django-cors-headers
djangorestframework-simplejwt
python-dotenv
pillow
psycopg2-binary
